requirements.txt
fastapi
uvicorn
python-dotenv
groq
scikit-learn
beautifulsoup4
requests
PyPDF2
jinja2
aiofiles
python-multipart==0.0.6
email-validator==2.1.0.post1
